refactor: move Firestore utility functions to server module and update imports
